client: extract execution module
The new `execution` module will host the shared execution logic for all
request types. `RequestExecutionParams` are now the main actor, who can
run any request with timeouts, retries, and speculative execution.

The `Session` is now responsible for constructing the
`RequestExecutionParams` only. Well, not entirely, it still has to
handle side effects such as schema or keyspace change.

Further refactoring is planned to migrate `QueryPager`'s execution logic
to the new module.

Important: this commit only moves whole blocks of code around. No logic
has been changed, and no intra-function refactorings were performed.
