+++ scylla/src/client/mod.rs
@@ -17,6 +17,8 @@
 
 pub mod execution_profile;
 
+mod execution;
+
 pub mod pager;
 
 pub mod client_routes;
